Automate scale test
Problem:
Non-functional scale test needs to be run manually.

Solution:
- Automate scale test.
- Use in-cluster Prometheus to collect CPU, memory and NGF metrics.
- Use Kubernetes API server to get NGF logs.

For development and troubleshooting, it is possible to run scale test
locally in Kind cluster. However, it is necessary to bring down
the number of HTTPRoutes to 50 or less (roughly).

Testing:
- Ran this test locally with 64 listeners, 50 routes and 50 upstreams.
- Ran this test on GKE with the default configuration.

Out of scope: ensuring this test runs successfully via GitHub pipeline.

Closes https://github.com/nginxinc/nginx-gateway-fabric/issues/1368

Largely based on work by Ciara in
https://github.com/nginxinc/nginx-gateway-fabric/pull/1804

Co-authored-by: Ciara Stacke <c.stacke@f5.com>